Add studentsAssigned field to AdminSchema and associated method
This field is used to keep track of the number of students assigned to
the coadmin. The schema method getNextCoadmin() is used to select the
coadmin to whom the next student will be assigned. We pick the coadmin
with the least value of studentsAssigned inorder to balance the load
among the coadmins.